Make registers 8bit instead of 16bit

+++ src/components/memory.h
@@ -10,7 +10,7 @@
 
 extern uint16_t program_counter; // Points at the current instruction in memory
 extern uint16_t index_register; // Points at locations in memory
-extern uint16_t registers[16]; //The CPU registers V0 - VF TODO: error handling for wrong access
+extern uint8_t registers[16]; //The CPU registers V0 - VF TODO: error handling for wrong access
 
 void init_memory();
 uint8_t memory_read(uint16_t address);
